Set up repository.
During IMO 2015 I analyzed various possible algorithms for determining
IMO medal boundaries, with a view to reducing the extent to which the
merits of particular candidate boundaries need to be debated by the
Jury each year (equally applicable to other mathematical olympiads
with similar rules on medal boundaries).  This repository will be used
for implementations of such algorithms, cleaned up for public release.